removed middleware and added role check
