+++ Portal/src/Datahub.Application/Services/ReverseProxy/IReverseProxyConfigService.cs
@@ -7,6 +7,9 @@ public interface IReverseProxyConfigService
     public const string WebAppPrefix = "webapp";
 
     public const string WorkspaceACLTransform = "ACLTransformProvider";
+    public const string WorkspaceRouteInfo = "Workspace";
+
+    public const string WorkspaceAuthorizationPolicy = "WorkspaceAppPolicy";
     ReverseProxyConfig GetConfigurationFromProjects();
 
     List<(string Acronym, RouteConfig Route, ClusterConfig Cluster)> GetAllConfigurationFromProjects();
