+++ Portal/src/Datahub.Portal/Datahub.Portal.csproj
@@ -44,4 +44,7 @@
 		<ProjectReference Include="..\Datahub.Metadata\Datahub.Metadata.csproj" />
 		<ProjectReference Include="..\Datahub.Portal.Metadata\Datahub.Portal.Metadata.csproj" />
 	</ItemGroup>
+	<ItemGroup>
+	  <Folder Include="Middleware\" />
+	</ItemGroup>
 </Project>
